fix: Random ENOENT errors in `copy_file_or_directory` on Windows
This happens for more or less the same files, usually on a batch update (eg. after typescript rebuilds a project and emits a lot of files). Looks like only `fs-extra` throws error, `fs` doesn't.

BTW, we don't need to use `pify` for `fse`, as `fs-extra` supports promises out of the box. But simply removing the `promisified.` prefix doesn't solve the issue.